feat: replaced new DocumentToolbar component in design-system
