--- packages/design-system/src/connect/components/document-toolbar-v2/index.ts
@@ -1,1 +1,0 @@
-export * from "./document-toolbar-v2.js";
